varios cambios principalmente, ya no se permite eliminar un equipo si tiene series anexas

+++ database/migrations/2020_03_17_155417_create_iccs_table.php
@@ -16,8 +16,9 @@ class CreateIccsTable extends Migration
         Schema::create('iccs', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->char('icc', 19)->unique();
-            $table->integer('sucursal_id');
-            $table->integer('status_id');   
+            $table->unsignedBigInteger('sucursal_id');
+            // $table->foreign('sucursal_id')->references('id')->on('sucursales');  
+            $table->integer('status_id'); 
             $table->timestamps();
             $table->softDeletes();
